+++ scripts/load-templates.js
